+++ packages/measure/src/types.ts
@@ -17,6 +17,9 @@ export interface MeasureResults {
   /** Array of measured render/execution durations for each run */
   durations: number[];
 
+  /** Array of measured render/execution durations for each warmup run   */
+  warmupDurations: number[];
+
   /** Arithmetic average of measured render/execution count for each run */
   meanCount: number;
 
@@ -25,4 +28,24 @@ export interface MeasureResults {
 
   /** Array of measured render/execution count for each run */
   counts: number[];
+}
+
+export interface MeasureRendersResults extends MeasureResults {
+  issues: RenderIssues;
+}
+
+export interface RenderIssues {
+  /**
+   * Update renders (re-renders) that happened immediately after component was created
+   * e.g., synchronous `useEffects` containing `setState`.
+   *
+   * This types of re-renders can be optimized by initializing the component with proper state in
+   * the initial render.
+   */
+  initialUpdateCount: number;
+
+  /**
+   * Re-renders that resulted in rendering the same output as the previous render. This arrays contains numbers of render
+   */
+  redundantUpdates: number[];
 }