+++ packages/reassure/src/bin/reassure.js
@@ -1,0 +1,7 @@
+#!/usr/bin/env node
+
+const importLocal = require('import-local');
+
+if (!importLocal(__filename)) {
+  require('@callstack/reassure-cli/bin/reassure');
+}