+++ packages/reassure/src/index.ts
@@ -1,4 +1,25 @@
-export { measurePerformance, measureFunction, configure, resetToDefault } from '@callstack/reassure-measure';
+export {
+  measureRenders,
+  measureFunction,
+  configure,
+  resetToDefaults,
+  measurePerformance,
+} from '@callstack/reassure-measure';
 export { dangerReassure } from '@callstack/reassure-danger';
 
-export type { MeasureOptions, MeasureFunctionOptions } from '@callstack/reassure-measure';
+export type {
+  MeasureResults,
+  MeasureRendersOptions,
+  MeasureFunctionOptions,
+  MeasureType,
+} from '@callstack/reassure-measure';
+export type {
+  MeasureHeader,
+  MeasureMetadata,
+  MeasureEntry,
+  CompareResult,
+  CompareMetadata,
+  CompareEntry,
+  AddedEntry,
+  RemovedEntry,
+} from '@callstack/reassure-compare';