+++ packages/reassure/CHANGELOG.md
@@ -1,5 +1,70 @@
 # reassure
 
+## 1.0.0-rc.5
+
+### Minor Changes
+
+- 04be1d4: Detect render issues (initial render updates, redundant renders)
+
+### Patch Changes
+
+- Updated dependencies [04be1d4]
+  - @callstack/reassure-compare@1.0.0-rc.5
+  - @callstack/reassure-measure@1.0.0-rc.5
+  - @callstack/reassure-cli@1.0.0-rc.5
+  - @callstack/reassure-danger@1.0.0-rc.5
+
+## 1.0.0-rc.4
+
+### Patch Changes
+
+- ea761f8: `reassure` enables WASM support by default and runs using v8 baseline compilers: sparkplug and liftoff (WASM)
+- Updated dependencies [ea761f8]
+  - @callstack/reassure-cli@1.0.0-rc.4
+  - @callstack/reassure-compare@1.0.0-rc.4
+  - @callstack/reassure-danger@1.0.0-rc.4
+  - @callstack/reassure-measure@1.0.0-rc.4
+
+## 1.0.0-rc.3
+
+### Minor Changes
+
+- 2ef5f4d: chore: fix version deps
+
+### Patch Changes
+
+- Updated dependencies [2ef5f4d]
+  - @callstack/reassure-measure@1.0.0-rc.3
+  - @callstack/reassure-danger@1.0.0-rc.3
+  - @callstack/reassure-cli@1.0.0-rc.3
+
+## 1.0.0-rc.2
+
+### Minor Changes
+
+- f722a6d: chore: migrate to Yarn Berry (4.x)
+
+### Patch Changes
+
+- Updated dependencies [f722a6d]
+  - @callstack/reassure-measure@1.0.0-rc.2
+  - @callstack/reassure-danger@1.0.0-rc.2
+  - @callstack/reassure-cli@1.0.0-rc.2
+
+## 1.0.0 RC
+
+### Major Changes
+
+- v1.0
+- 4352279: - Rename `measurePerformance` to `measureRenders`, `resetToDefault` to `resetToDefaults`.
+
+### Patch Changes
+
+- Updated dependencies [f9116d0]
+  - @callstack/reassure-cli@1.0.0-rc.1
+  - @callstack/reassure-danger@1.0.0-rc.1
+  - @callstack/reassure-measure@1.0.0-rc.1
+
 ## 0.11.0
 
 ### Minor Changes